+++ final/FinalProject/Stock.cs
@@ -3,8 +3,8 @@ class Stock
     Ingredient _ingredient;
     int _availableQty;
     string _uoM;
-    decimal _unitCost;
-    public Stock(Ingredient ingredient, int quantity, string unitOfMeasure, decimal cost)
+    float _unitCost;
+    public Stock(Ingredient ingredient, int quantity, string unitOfMeasure, float cost)
     {
         _ingredient =ingredient;
         _availableQty = quantity;
@@ -15,7 +15,7 @@ class Stock
     {
         return true;
     }
-    public void UpdateElement(int quantity, string unitOfMeasure, decimal cost)
+    public void UpdateElement(int quantity, string unitOfMeasure, float cost)
     {
         
     }
